requirements.txt
amqp==2.5.2
aniso8601==8.0.0
billiard==3.5.0.5
blis==0.4.1
boto==2.49.0
boto3==1.13.11
botocore==1.16.11
catalogue==1.0.0
celery==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
console-log==0.2.10
cymem==2.0.3
Cython==0.29.14
dic==1.5.2b1
dnspython==1.16.0
docutils==0.15.2
eventlet==0.25.2
Flask==1.1.2
Flask-RESTful==0.3.8
flower==0.9.4
gensim==3.8.3
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
kombu==4.6.8
MarkupSafe==1.1.1
monotonic==1.5
murmurhash==1.0.2
nltk==3.5
numpy==1.18.4
pandas==1.0.3
pika==1.1.0
plac==1.1.3
preshed==3.0.2
pycparser==2.20
pymongo==3.10.1
python-dateutil==2.8.1
python-rake==1.4.5
pytz==2020.1
rake-nltk==1.0.4
regex==2020.5.14
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==2.0.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
tornado==6.0.4
tqdm==4.46.0
urllib3==1.25.9
vine==1.3.0
wasabi==0.6.0
Werkzeug==1.0.1
wsgigzip==0.1.4
zipp==3.1.0
